+++ include/SimplicialManifold.h
@@ -54,7 +54,7 @@ using Geometry_tuple =
 /// This function iterates over all edges in the triangulation
 /// and classifies them as timelike or spacelike.
 /// Timelike edges are stored in the **timelike_edges** vector as an Edge_handle
-/// (tuple of Cell_handle, std::intmax_t, std::intmax_t) for later use by
+/// (tuple of Cell_handle, std::int_fast32_t, std::int_fast32_t) for later use by
 /// ergodic moves on timelike edges. Spacelike edges are also stored as a
 /// vector of Edge_handle **spacelike_edges**, for use by (4,4) moves as
 /// well as the distance-finding algorithms.
@@ -82,8 +82,8 @@ auto classify_edges(T&& universe_ptr)
 
     // Make Edge_handle
     Edge_handle thisEdge{
-        ch, static_cast<std::intmax_t>(ch->index(ch->vertex(eit->second))),
-        static_cast<std::intmax_t>(ch->index(ch->vertex(eit->third)))};
+        ch, static_cast<std::int_fast32_t>(ch->index(ch->vertex(eit->second))),
+        static_cast<std::int_fast32_t>(ch->index(ch->vertex(eit->third)))};
 
     if (time1 != time2) {  // We have a timelike edge
       timelike_edges.emplace_back(thisEdge);
@@ -139,16 +139,16 @@ auto classify_simplices(T&& universe_ptr)
   for (cit = universe_ptr->finite_cells_begin();
        cit != universe_ptr->finite_cells_end(); ++cit)
   {
-    std::intmax_t max_values{0};
-    std::intmax_t min_values{0};
+    std::int_fast32_t max_values{0};
+    std::int_fast32_t min_values{0};
     // Push every time value of every vertex into a list
-    std::intmax_t timevalues[4] = {
+    std::int_fast32_t timevalues[4] = {
         cit->vertex(0)->info(),
         cit->vertex(1)->info(),
         cit->vertex(2)->info(),
         cit->vertex(3)->info(),
     };
-    std::intmax_t max_time =
+    std::int_fast32_t max_time =
         *std::max_element(std::begin(timevalues), std::end(timevalues));
     for (auto elt : timevalues) {
       if (elt == max_time) {
@@ -244,15 +244,15 @@ struct GeometryInfo
   std::vector<Vertex_handle> vertices;
 
   /// @brief Spacelike facets for each timeslice
-  boost::optional<std::multimap<intmax_t, Facet>> spacelike_facets;
+  boost::optional<std::multimap<int_fast32_t, Facet>> spacelike_facets;
 
   /// @brief Actual timevalues of simulation
-  boost::optional<std::set<intmax_t>> timevalues;
+  boost::optional<std::set<int_fast32_t>> timevalues;
 
  public:
   /// @brief Getter for spacelike facets
   /// @return The multimap of facets
-  const boost::optional<std::multimap<intmax_t, Facet>>& getSpacelike_facets()
+  const boost::optional<std::multimap<int_fast32_t, Facet>>& getSpacelike_facets()
       const
   {
     return spacelike_facets;
@@ -261,21 +261,21 @@ struct GeometryInfo
   /// @brief Setter for spacelike facets
   /// @param spacelike_facets The multimap of facets
   void setSpacelike_facets(
-      const boost::optional<std::multimap<intmax_t, Facet>>& spacelike_facets)
+      const boost::optional<std::multimap<int_fast32_t, Facet>>& spacelike_facets)
   {
     GeometryInfo::spacelike_facets = spacelike_facets;
   }
 
   /// @brief Getter for timevalues
   /// @return The set of timevalues
-  const boost::optional<std::set<intmax_t>>& getTimevalues() const
+  const boost::optional<std::set<int_fast32_t>>& getTimevalues() const
   {
     return timevalues;
   }
 
   /// @brief Setter for timevalues
   /// @param timevalues The set of timevalues
-  void setTimevalues(const boost::optional<std::set<intmax_t>>& timevalues)
+  void setTimevalues(const boost::optional<std::set<int_fast32_t>>& timevalues)
   {
     GeometryInfo::timevalues = timevalues;
   }
@@ -328,21 +328,21 @@ struct GeometryInfo
 
   /// @brief Timelike edges
   /// @return The number of edges spanning timeslices
-  auto N1_TL() { return static_cast<std::uintmax_t>(timelike_edges.size()); }
+  auto N1_TL() { return static_cast<std::uint_fast32_t>(timelike_edges.size()); }
 
   /// @brief Spacelike edges
   /// @return The number of edges on same timeslice
-  auto N1_SL() { return static_cast<std::uintmax_t>(spacelike_edges.size()); }
+  auto N1_SL() { return static_cast<std::uint_fast32_t>(spacelike_edges.size()); }
 
   /// @brief (3,1) simplices
   /// @return The total number of simplices with 3 vertices on the t
   /// timeslice and 1 vertex on the t+1 timeslice
-  auto N3_31() { return static_cast<std::uintmax_t>(three_one.size()); }
+  auto N3_31() { return static_cast<std::uint_fast32_t>(three_one.size()); }
 
   /// @brief (1,3) simplices
   /// @return The total number of simplices with 1 vertex on the t timeslice
   /// and 3 vertices on the t+1 timeslice
-  auto N3_13() { return static_cast<std::uintmax_t>(one_three.size()); }
+  auto N3_13() { return static_cast<std::uint_fast32_t>(one_three.size()); }
 
   /// @brief (3,1) and (1,3) simplices
   /// @return The total number of simplices with 3 vertices on one
@@ -354,7 +354,7 @@ struct GeometryInfo
   /// @return The total number of simplices with 2 vertices on one
   /// timeslice and 2 vertices on the adjacent timeslice. Used to
   /// calculate the change in action.
-  auto N3_22() { return static_cast<std::uintmax_t>(two_two.size()); }
+  auto N3_22() { return static_cast<std::uint_fast32_t>(two_two.size()); }
 
   /// @brief Number of cells
   ///
@@ -373,18 +373,18 @@ struct GeometryInfo
   auto number_of_edges() { return N1_TL() + N1_SL(); }
 
   //  auto max_timevalue() { return *timevalues.crbegin();}
-  boost::optional<std::intmax_t> max_timevalue()
+  boost::optional<std::int_fast32_t> max_timevalue()
   {
     return timevalues ? *timevalues->crbegin() : 0;
   }
 
-  boost::optional<std::intmax_t> min_timevalue()
+  boost::optional<std::int_fast32_t> min_timevalue()
   {
     return timevalues ? *timevalues->begin() : 0;
   }
   /// @brief Number of vertices
   /// @return The number of vertices in the triangulation
-  auto N0() { return static_cast<std::uintmax_t>(vertices.size()); }
+  auto N0() { return static_cast<std::uint_fast32_t>(vertices.size()); }
 
   template <typename T1, typename T2>
   friend auto make_23_move(T1&& universe, T2&& attempted_moves)
@@ -463,7 +463,7 @@ struct SimplicialManifold
   /// @param simplices The number of desired simplices in the triangulation
   /// @param timeslices The number of timeslices in the triangulation
   /// @return A populated SimplicialManifold{}
-  SimplicialManifold(std::intmax_t simplices, std::intmax_t timeslices)
+  SimplicialManifold(std::int_fast32_t simplices, std::int_fast32_t timeslices)
       : triangulation{make_triangulation(simplices, timeslices)}
       , geometry{std::make_unique<GeometryInfo>(
             classify_all_simplices(triangulation))}
